Move plex into apps
