+++ pkg/configfile/config.go
@@ -18,9 +18,9 @@ import (
 
 	"github.com/BurntSushi/toml"
 	"github.com/Notifiarr/notifiarr/pkg/apps"
+	"github.com/Notifiarr/notifiarr/pkg/apps/apppkg/plex"
 	"github.com/Notifiarr/notifiarr/pkg/logs"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
-	"github.com/Notifiarr/notifiarr/pkg/plex"
 	"github.com/Notifiarr/notifiarr/pkg/services"
 	"github.com/Notifiarr/notifiarr/pkg/snapshot"
 	"github.com/Notifiarr/notifiarr/pkg/triggers"
@@ -56,7 +56,6 @@ type Config struct {
 	Timeout    cnfg.Duration          `json:"timeout" toml:"timeout" xml:"timeout" yaml:"timeout"`
 	Serial     bool                   `json:"serial" toml:"serial" xml:"serial" yaml:"serial"`
 	Retries    int                    `json:"retries" toml:"retries" xml:"retries" yaml:"retries"`
-	Plex       *plex.Server           `json:"plex" toml:"plex" xml:"plex" yaml:"plex"`
 	Snapshot   *snapshot.Config       `json:"snapshot" toml:"snapshot" xml:"snapshot" yaml:"snapshot"`
 	Services   *services.Config       `json:"services" toml:"services" xml:"services" yaml:"services"`
 	Service    []*services.Service    `json:"service" toml:"service" xml:"service" yaml:"service"`
@@ -159,7 +158,6 @@ func (c *Config) Get(flag *Flags) (*website.Server, *triggers.Actions, error) {
 	// This config contains [some of] the same data as the normal Config.
 	c.Services.Website = website.New(&website.Config{
 		Apps:    c.Apps,
-		Plex:    c.Plex,
 		Logger:  c.Services.Logger,
 		BaseURL: website.BaseURL,
 		Timeout: c.Timeout,
@@ -209,7 +207,6 @@ func (c *Config) setup() *triggers.Actions {
 
 	return triggers.New(&triggers.Config{
 		Apps:       c.Apps,
-		Plex:       c.Plex,
 		Serial:     c.Serial,
 		Website:    c.Services.Website,
 		Snapshot:   c.Snapshot,
