+++ pkg/triggers/triggers.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
-	"github.com/Notifiarr/notifiarr/pkg/plex"
 	"github.com/Notifiarr/notifiarr/pkg/snapshot"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/backups"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/cfsync"
@@ -28,7 +27,6 @@ import (
 type Config struct {
 	Serial     bool
 	Apps       *apps.Apps
-	Plex       *plex.Server
 	Website    *website.Server
 	Snapshot   *snapshot.Config
 	WatchFiles []*filewatch.WatchFile
@@ -62,7 +60,7 @@ func New(config *Config) *Actions {
 		Serial:   config.Serial,
 		Logger:   config.Logger,
 	}
-	plex := plexcron.New(common, config.Plex)
+	plex := plexcron.New(common, config.Apps.Plex)
 
 	return &Actions{
 		PlexCron:   plex,
