+++ pkg/website/server.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
-	"github.com/Notifiarr/notifiarr/pkg/plex"
 	"github.com/shirou/gopsutil/v3/host"
 	"golift.io/cnfg"
 )
@@ -37,7 +36,6 @@ var (
 // Config is the input data needed to send payloads to notifiarr.
 type Config struct {
 	Apps       *apps.Apps
-	Plex       *plex.Server // plex sessions
 	Serial     bool
 	Retries    int
 	BaseURL    string
